test: focus shared ui integration contracts
